Create adequate canvas automatically
This makes non-square non-power-of-two images into square power-of-two
images by adding adequate white surroundings. With this, the program can
now take input images of any dimensions.

In the future one could improve the algorithms to work with less strict
requirements, but this is more work, more error-prone and the program is
really fast enough that the extra canvas space should not matter.

fixes #1